requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0

pip install numpy==1.24.3 --user
pip install pillow==10.0.0 --user
pip install flask==2.3.3 --user
pip install flask-cors==4.0.0 --user
pip install opencv-python==4.8.1.78 --user
pip install python-dotenv==1.0.0 --user
pip install tensorflow==2.16.2

pip install flask==2.3.3
pip install flask-cors==4.0.0
pip install tensorflow==2.13.0
pip install pillow==10.0.0
pip install numpy==1.24.3
pip install opencv-python==4.8.1.78
pip install python-dotenv==1.0.0
pip install opencv-python-headless==4.10.0.84
